project_template/requirements.txt
appdirs==1.4.4            # via black, virtualenv
asgiref==3.2.10           # via django
black==20.8b1             # via -r requirements.in
certifi==2020.6.20        # via requests
cfgv==3.2.0               # via pre-commit
chardet==3.0.4            # via requests
click==7.1.2              # via black, pip-tools
coreapi==2.3.3            # via -r requirements.in, drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
distlib==0.3.1            # via virtualenv
django-ipware==3.0.1      # via -r requirements.in
django==3.0.5             # via -r requirements.in, djangorestframework, drf-yasg
djangorestframework==3.11.1  # via -r requirements.in, drf-yasg
drf-yasg==1.17.1          # via -r requirements.in
filelock==3.0.12          # via virtualenv
gunicorn==20.0.4          # via -r requirements.in
identify==1.4.30          # via pre-commit
idna==2.10                # via requests
importlib-metadata==1.7.0  # via pre-commit, virtualenv
inflection==0.5.1         # via drf-yasg
itypes==1.2.0             # via coreapi
jinja2==2.11.2            # via coreschema
markupsafe==1.1.1         # via jinja2
mypy-extensions==0.4.3    # via black
nodeenv==1.5.0            # via pre-commit
packaging==20.4           # via drf-yasg
pathspec==0.8.0           # via black
pip-tools==5.3.1          # via -r requirements.in
pre-commit==2.7.1         # via -r requirements.in
psycopg2-binary==2.8.6    # via -r requirements.in
pyparsing==2.4.7          # via packaging
python-dotenv==0.14.0     # via -r requirements.in
pytz==2020.1              # via django
pyyaml==5.3.1             # via pre-commit
regex==2020.7.14          # via black
requests==2.24.0          # via coreapi
ruamel.yaml.clib==0.2.2   # via ruamel.yaml
ruamel.yaml==0.16.12      # via drf-yasg
six==1.15.0               # via drf-yasg, packaging, pip-tools, virtualenv
sqlparse==0.2.4           # via django,
toml==0.10.1              # via black, pre-commit
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
uritemplate==3.0.1        # via coreapi, drf-yasg
urllib3==1.25.10          # via requests
virtualenv==20.0.31       # via pre-commit
zipp==3.1.0               # via importlib-metadata
django-cors-headers==3.5.0

